requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
pandas-ta>=0.3.14b
python-dotenv>=1.0.0
APScheduler>=3.10.0
python-telegram-bot>=20.0
mplfinance>=0.12.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
numpy>=1.24.0
